Separated the facotory tests into a .test file
