+++ src/Transformation/Transformation.test.ts
@@ -3,19 +3,6 @@ import { MatrixSizeError } from '../Errors/MatrixSizeError';
 import { TransformationConcreteProduct } from '../Mocks/TransformationConcreteProduct';
 import { TransformationFactory } from './TransformationFactory';
 
-describe('Transformation factory', () => {
-  it('should not be throwing any error on creating a transformation creator', () => {
-    expect(TransformationFactory.createTransformationCreator).not.toThrowError();
-  });
-
-  it('should be instancing the tranformation creator', () => {
-    const tranformationCreator = TransformationFactory.createTransformationCreator();
-
-    expect(tranformationCreator).toBeDefined();
-    expect(tranformationCreator.apply).toBeDefined();
-  });
-});
-
 describe('Transformation class', () => {
   it('should not be throwing error when creating your respective product', () => {
     const tranformationCreator = TransformationFactory.createTransformationCreator();
