--- src/Transformation/TransformationMatrixContract.ts
@@ -1,3 +1,0 @@
-export interface TransformationMatrixContract {
-  assignRate(): void
-}
